Reorganize files and simple tests
Reorganize files with long term maintainability in mind. Also further
simplify existing tests and fix some busted ones.
