+++ server/api/routes/generators.routes.js
@@ -1,6 +1,6 @@
 import {
   generateWeather,
-} from '../controllers/generators.controller';
+} from '../controllers/generators/weather.controller';
 
 module.exports = (router) => {
   router.route('/api/generators/weather').post(generateWeather);
